Supports I/O in char/byte during transformation operations.
Transformers can now read and write data in chars and not only in bytes anymore. This can reduce the number of byte/char conversions if two successive transformers are manipulating char stream. The API has been completely refactored to manipulate an abstraction of I/O operations called Input and Output that can respectively read and write data with a preferred approach (byte or char). This change has introduced new challenges around char <-> byte conversion that require for a lot of internal objects created on the fly to use a particular charset. This issue has been taken has the opportunity to let the engines and DAO be configured with a contextualized charset resolved from the properties. The charset is now embedded by the NutType that has moved from an enum to class status in order to be instantiated with a particular charset. NutType are consequently now produced by a NutTypeFactory that knows the charset to use and that is created by the ContextBuilder and injected to the components via an inspector.

Fixes #183