Adds a filter that translates 'node_module' to 'webjars' URL.
A new filter mapped to the /node_modules/* URL pattern is installed when the WUIC support for WebJar is added to the classpath. This filter will intercept any classic 'node_modules' URL obtain with the 'npm install' command and redirect it to the corresponding 'webjars' URL which is valid as soon as the underlying dependency is added to the classpath also.

Fixes https://github.com/wuic/wuic/issues/243